+++ FO-DICOM.Core/Network/DesktopNetworkListener.cs
@@ -63,6 +63,8 @@ namespace FellowOakDicom.Network
         public async Task<INetworkStream> AcceptNetworkStreamAsync(
             ITlsAcceptor tlsAcceptor,
             bool noDelay,
+            int? receiveBufferSize,
+            int? sendBufferSize,
             ILogger logger,
             CancellationToken token)
         {
@@ -82,6 +84,14 @@ namespace FellowOakDicom.Network
                 {
                     var tcpClient = await acceptTcpClientTask;
                     tcpClient.NoDelay = noDelay;
+                    if (receiveBufferSize.HasValue)
+                    {
+                        tcpClient.ReceiveBufferSize = receiveBufferSize.Value;
+                    }
+                    if (sendBufferSize.HasValue)
+                    {
+                        tcpClient.SendBufferSize = sendBufferSize.Value;
+                    }
                     
                     if (logger.IsEnabled(LogLevel.Debug))
                     {