+++ ChangeLog.md
@@ -2,6 +2,7 @@
 * Fix issue where extracting a string from a DICOM dataset could return null if the tag was present but empty
 * Immediately throw an exception if DICOM server synchronously fails to start (#1562)
 * Fix issue where stopping a DICOM server left the unused services cleanup task running (#1562)
+* Add the possibility to configure TCP buffer sizes (#1564)
 * Fix incorrect values returned from `DicomEncoding.GetCharset()` (#1624)
 * Tolerate `Specific Character Set` values misspelled as "ISO-IR ###" additionally to "ISO IR ###"